Do not remove _i18n property as it can be needed after beforeDestroy()
In some circumstances a long running method call can still be running after a component is removed from the DOM. This causes beforeDestroy() to be called.

If this long running method now calls this.$t or similar that leads to errors as the further down called this._i18n is null. This._i18n needs to be available even after beforeDestroy() is being called.

fixes #184