Anchor regexp to beginning of filename.
This regular expression is trying to find files of the form "foo.json",
where "foo" doesn't contain any special characters.  But the regex is
used as a substring match, so it needs to anchor to the beginning of the
basename, otherwise it will match anything which ends with "x.json".