Add onComponentInstanceCreated constructor option
Takes listener function that will be called when component-local
instance of VueI18n is created.

This allows for extending component-local VueI18n instances. Those are
created when Vue component specifies i18n options in an "i18n" property
as a plain object.

This allows Nuxt to also extend those instances in addition to extending
the root instance, so that there are no missing properties that would
cause users code to crash.

Fixes #887